requirements.txt
torch
torchvision
pytorch-lightning
hydra-core
nltk  # For BLEU and WER metrics